requirements.txt
Keras==2.2.4
tensorflow==1.9.0
numpy==1.16.2
scikit-learn==0.20.3
matplotlib==3.0.3
pytest==4.4.0
PyYAML==5.1